requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # file uploads
httpx==0.27.0                    # async HTTP client

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0                  # async PostgreSQL driver
alembic==1.13.1                  # DB migrations
# psycopg2-binary removed - not needed on Windows; asyncpg handles async ops

# ─── Caching ─────────────────────────────────────────────────────────────────
redis==5.0.4

# ─── Vector DB ───────────────────────────────────────────────────────────────
qdrant-client==1.9.1

# ─── Auth & Security ─────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7
python-dotenv==1.0.1

# ─── AI / LLM ────────────────────────────────────────────────────────────────
anthropic==0.28.0                # Claude API (primary LLM)
google-generativeai==0.7.2       # Gemini (fallback LLM)
openai==1.30.5                   # optional fallback

# ─── PII Redaction ───────────────────────────────────────────────────────────
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4                     # required by presidio

# ─── NLP / Medical NER ───────────────────────────────────────────────────────
# Run after install: python -m spacy download en_core_web_lg
# For medical NER (optional, heavier): pip install scispacy
# then: pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz

# ─── Sentence Embeddings ─────────────────────────────────────────────────────
sentence-transformers==3.0.1     # BioLORD-2023 or PubMedBERT embeddings
torch==2.3.0                     # required by sentence-transformers
transformers==4.41.2

# ─── Image Processing ────────────────────────────────────────────────────────
Pillow==10.3.0
python-magic==0.4.27

# ─── File Storage (MinIO / S3-compatible) ────────────────────────────────────
minio==7.2.7
boto3==1.34.105                  # alternative: AWS S3

# ─── Background Tasks ────────────────────────────────────────────────────────
celery==5.4.0
celery[redis]==5.4.0
flower==2.0.1                    # Celery monitoring UI

# ─── PubMed / Research ───────────────────────────────────────────────────────
biopython==1.83                  # PubMed E-utilities via Entrez
requests==2.32.2
beautifulsoup4==4.12.3
lxml==5.2.1

# ─── Data Validation ─────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── Monitoring & Logging ────────────────────────────────────────────────────
loguru==0.7.2
prometheus-fastapi-instrumentator==7.0.0

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
factory-boy==3.3.0
